Extract grouped query decoder
